+++ myems-web/src/components/MyEMS/CombinedEquipment/CombinedEquipmentEfficiency.js
@@ -76,7 +76,7 @@ const CombinedEquipmentEfficiency = ({ setRedirect, setRedirectUrl, t }) => {
   const [selectedSpaceName, setSelectedSpaceName] = useState(undefined);
   const [combinedEquipmentList, setCombinedEquipmentList] = useState([]);
   const [selectedCombinedEquipment, setSelectedCombinedEquipment] = useState(undefined);
-  const [comparisonType, setComparisonType] = useState('month-on-month');
+  const [comparisonType, setComparisonType] = useState('none-comparison');
   const [periodType, setPeriodType] = useState('daily');
   const [cascaderOptions, setCascaderOptions] = useState(undefined);
   const [basePeriodDateRange, setBasePeriodDateRange] = useState([
@@ -1474,7 +1474,7 @@ const CombinedEquipmentEfficiency = ({ setRedirect, setRedirectUrl, t }) => {
                     id="comparisonType"
                     name="comparisonType"
                     bsSize="sm"
-                    defaultValue="month-on-month"
+                    defaultValue="none-comparison"
                     onChange={onComparisonTypeChange}
                   >
                     {comparisonTypeOptions.map((comparisonType, index) => (