+++ myems-web/src/components/MyEMS/Microgrid/MicrogridReportingEnergy.js
@@ -72,7 +72,7 @@ const MicrogridReportingEnergy = ({ setRedirect, setRedirectUrl, t }) => {
   const [microgridList, setMicrogridList] = useState([]);
   const [filteredMicrogridList, setFilteredMicrogridList] = useState([]);
   const [selectedMicrogrid, setSelectedMicrogrid] = useState(undefined);
-  const [comparisonType, setComparisonType] = useState('month-on-month');
+  const [comparisonType, setComparisonType] = useState('none-comparison');
   const [periodType, setPeriodType] = useState('daily');
   const [cascaderOptions, setCascaderOptions] = useState(undefined);
   const [basePeriodDateRange, setBasePeriodDateRange] = useState([
@@ -826,7 +826,7 @@ const MicrogridReportingEnergy = ({ setRedirect, setRedirectUrl, t }) => {
                     id="comparisonType"
                     name="comparisonType"
                     bsSize="sm"
-                    defaultValue="month-on-month"
+                    defaultValue="none-comparison"
                     onChange={onComparisonTypeChange}
                   >
                     {comparisonTypeOptions.map((comparisonType, index) => (